Migrate campaign endpoints to new pattern
